Include Python 3 recordtype until version is released
There is a PR open at https://bitbucket.org/ericvsmith/recordtype/pull-requests/1/add-python-3-support/diff. Until that is merged and recordtype 1.2 is released, I just brought this into this app. The recordtype repo is mercurial not git so I didn't see a clear way to install that Python 3 branch. Decided to import the file here since recordtype lib is just a single file.